requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
openpyxl>=3.1
django-environ>=0.11
